+++ app/composables/render/update/aircraft.ts
@@ -37,9 +37,11 @@ export function updateAircraft(context: DataUpdateContext) {
 
     for (const aircraft of dataStore.vatsim.data.pilots.value) {
         if (aircraft.departure) addAircraftToAirport(context, aircraft, aircraft.departure, (aircraft.departure === aircraft.airport && aircraft.status === 'depTaxi') ? 'groundDep' : 'departures');
+        else if (aircraft.airport) addAircraftToAirport(context, aircraft, aircraft.airport, aircraft.status === 'depTaxi' ? 'groundDep' : 'groundArr');
+
         if (aircraft.arrival) addAircraftToAirport(context, aircraft, aircraft.arrival, aircraft.status === 'arrTaxi' ? 'groundArr' : 'arrivals');
 
-        if (aircraft.airport && aircraft.airport !== aircraft.departure && aircraft.airport !== aircraft.arrival) {
+        if (aircraft.airport && aircraft.departure && aircraft.airport !== aircraft.departure && aircraft.airport !== aircraft.arrival) {
             addAircraftToAirport(context, aircraft, aircraft.airport, 'groundArr');
         }
     }